Implemented get(long position) method to get value at position. Added a test.

+++ main/java/datastructure/exceptions/NoElementAtPositionException.java
@@ -6,7 +6,7 @@ public class NoElementAtPositionException extends RuntimeException {
         super("There is no element at the position in list.");
     }
 
-    public NoElementAtPositionException(int position, int size) {
+    public NoElementAtPositionException(long position, long size) {
         super(String.format("There is no element at the position %d in list with size %d.", position, size));
     }
 }
